Update COSC344 ASSIGNMENT 2.md
Nat's normalization. As far as I can tell, most of my relations were already in BCNF (but that could just be wishful thinking; please correct me if I'm wrong!). The only exception was the Course entity, which had an undergraduate attribute and a postgraduate attribute. Assuming that a course must be either for undergraduates or postgraduates (and not both), there seems to be a transitive dependency (i.e., if a course changes the boolean value for its undergraduate attribute, the boolean value for its postgraduate attribute must also change). However, I'm not sure if the recommended step (removing transitive dependencies into a new relation) is appropriate here. I'd prefer to revise the ER diagram in step 1 of the assignment so that course has a "level" attribute instead of an undergraduate and a postgraduate attribute. Maybe we can discuss this later in the meeting?