Update university.sql
Adding sql code for department and course. (I'm not sold on course as I currently have it - what happens if a user wants to know how many years a course will take if it's completed on a part-time basis? - but perhaps the best time to tweak our database structure will be after everyone has submitted a rough version of their entities/relationships.